+++ CHANGELOG.md
@@ -7,8 +7,12 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
+* Initialized Repo
+* Setup `setup.py`
+* Added Fabric Commands
+* Setup changelog management
 
 ### Changes
-
-### Fixes
+* Moved `check.py` to `webmon/main.py`
+* Renamed `check_urls()` to `run()`